APIlark: New `apilark_api()` and `apilark_api_struct()` rule factories
These rule factories create "API rules", which are instantiated by the
API owner in their `BUILD` file.

IMPORTANT: These API rules are not actually useful yet, because there is
no way to depend on them at the moment.

However, this includes the full suite of functionality and covers most
of it with Starlark `analysis_test()` rules.