APIlark: Create internal `inject.bzl` library and tests
We need to be able to pass alternate definitions of `fail()` or other
builtins when testing some public APIs, but we don't actually want to
allow other users to do so.

Therefore, all injected values are wrapped by this `inject.bzl` library
to ensure they can only be passed by tests.