Save the `ImplContextInfo` provider to a global to avoid Bazel crashes
Each Starlark `provider()` class *MUST* be assigned as a global in the
creating `.bzl` file, per the Bazel documentation here:
https://bazel.build/rules/lib/globals/bzl.html#provider

Not doing so causes Bazel crashes in certain situations today, and will
become a hard error in some future version of Bazel.

Let's store it into a global variable to comply with the `.bzl` spec.