Replace non-exported `provider()` token in `apilark_key()` with a lambda.
Each Starlark `provider()` class *MUST* be assigned as a global in the
creating `.bzl` file, per the Bazel documentation here:
https://bazel.build/rules/lib/globals/bzl.html#provider

Not doing so causes Bazel crashes in certain situations today, and will
become a hard error in some future version of Bazel.

We can achieve the same result as the current `provider()` token with
the use of a `lambda: ...` instead; it will be globally-unique without
violating the `provider()` spec.